fix: #125 CP 검색 카테고리 분리
