+++ src/main/java/mediHub_be/cp/repository/CpSearchCategoryDataRepository.java
@@ -8,6 +8,7 @@ import org.springframework.data.repository.query.Param;
 import org.springframework.stereotype.Repository;
 
 import java.util.List;
+import java.util.Optional;
 
 @Repository
 public interface CpSearchCategoryDataRepository extends JpaRepository<CpSearchCategoryData, Long> {
@@ -44,7 +45,7 @@ public interface CpSearchCategoryDataRepository extends JpaRepository<CpSearchCa
             "LEFT JOIN CpSearchCategory AS csc ON cscd.cpSearchCategorySeq = csc.cpSearchCategorySeq " +
             "JOIN User AS u ON cscd.userSeq = u.userSeq " +
             "WHERE cscd.cpSearchCategoryDataSeq = :cpSearchCategoryDataSeq")
-    ResponseCpSearchCategoryDataDTO findByCpSearchCategoryDataSeq(@Param("cpSearchCategoryDataSeq") long cpSearchCategoryDataSeq);
+    Optional<ResponseCpSearchCategoryDataDTO> findByCpSearchCategoryDataSeq(@Param("cpSearchCategoryDataSeq") long cpSearchCategoryDataSeq);
 
     boolean existsByCpSearchCategoryDataName(String cpSearchCategoryDataName);
