+++ src/main/java/mediHub_be/cp/service/CpSearchCategoryService.java
@@ -5,9 +5,7 @@ import lombok.extern.slf4j.Slf4j;
 import mediHub_be.common.exception.CustomException;
 import mediHub_be.common.exception.ErrorCode;
 import mediHub_be.cp.dto.ResponseCpSearchCategoryDTO;
-import mediHub_be.cp.dto.ResponseCpSearchCategoryDataDTO;
 import mediHub_be.cp.entity.CpSearchCategory;
-import mediHub_be.cp.entity.CpSearchCategoryData;
 import mediHub_be.cp.repository.CpSearchCategoryDataRepository;
 import mediHub_be.cp.repository.CpSearchCategoryRepository;
 import mediHub_be.security.util.SecurityUtil;
@@ -30,33 +28,28 @@ public class CpSearchCategoryService {
     private final Logger logger = LoggerFactory.getLogger("mediHub_be.cp.service.CpSearchCategoryService");     // Logger
 
     /**
-     * 모든 CP 검색 카테고리를 조회하여 리스트로 반환합니다.
+     * CP 검색 카테고리 목록을 조회하는 서비스 메서드입니다.
      *
-     * @return CP 검색 카테고리 DTO 리스트
-     * @throws CustomException NOT_FOUND_CP_SEARCH_CATEGORY:
-     *                         조회 결과가 없을 경우 발생.
+     * <p>이 메서드는 데이터베이스에서 모든 CP 검색 카테고리를 조회하여
+     * 해당 카테고리의 정보를 리스트로 반환합니다.</p>
+     *
+     * @return ResponseCpSearchCategoryDTO 리스트, 카테고리가 존재하지 않을 경우 예외 발생
+     * @throws CustomException 데이터베이스 접근 오류 또는 카테고리가 발견되지 않은 경우 예외 발생
      */
     public List<ResponseCpSearchCategoryDTO> getCpSearchCategoryList() {
         logger.info("CP 검색 카테고리 조회 요청 시작");
 
-        List<ResponseCpSearchCategoryDTO> dtoList;
+        List<ResponseCpSearchCategoryDTO> dtoList = null;
 
         // DB 조회
         try {
             dtoList = cpSearchCategoryRepository.findJoinUserOnUserSeq();
-
-            if (dtoList.isEmpty()) {
-                logger.warn("CP 검색 카테고리 조회 결과가 비어 있습니다.");
-                throw new CustomException(ErrorCode.NOT_FOUND_CP_SEARCH_CATEGORY);
-            }
-
-            logger.info("CP 검색 카테고리 조회 성공: {} 카테고리", dtoList.size());
         } catch (DataAccessException e) {
             logger.error("데이터 접근 오류 발생: {}", e.getMessage(), e);
             throw new CustomException(ErrorCode.INTERNAL_DATA_ACCESS_ERROR);
         } catch (Exception e) {
             logger.error("예기치 않은 오류 발생: {}", e.getMessage(), e);
-            throw new CustomException(ErrorCode.INTERNAL_SERVER_ERROR);
+            throw new RuntimeException("CP 검색 카테고리 리스트 조회 과정에서 예상치 못한 에러가 발생했습니다.", e);
         }
 
         if (dtoList.isEmpty()) {
@@ -69,72 +62,86 @@ public class CpSearchCategoryService {
     }
 
     /**
-     * CP 검색 카테고리를 조회하는 서비스 메서드.
+     * 주어진 CP 검색 카테고리의 고유 번호를 사용하여 해당 카테고리를 조회하는 서비스 메서드입니다.
+     *
+     * <p>이 메서드는 CP 검색 카테고리 번호를 기반으로 데이터베이스에서 카테고리 정보를 조회하며,
+     * 카테고리가 존재하지 않거나 데이터 접근 중 오류가 발생하면 예외를 던집니다.</p>
      *
      * @param cpSearchCategorySeq 조회할 CP 검색 카테고리의 고유 번호
      * @return ResponseCpSearchCategoryDTO 해당 카테고리의 정보
-     * @throws CustomException 내부 데이터 접근 오류 또는 카테고리 미발견 시 예외 발생
+     * @throws CustomException 내부 데이터 접근 오류 또는 카테고리가 발견되지 않은 경우 예외 발생
      */
-    public ResponseCpSearchCategoryDTO getCpSearchCategory(long cpSearchCategorySeq) {
+    public ResponseCpSearchCategoryDTO getCpSearchCategoryByCpSearchCategorySeq(long cpSearchCategorySeq) {
         logger.info("CP 검색 카테고리 조회 시작: cpSearchCategorySeq={}", cpSearchCategorySeq);
 
-        ResponseCpSearchCategoryDTO dto = null;
-
         try {
-            dto = cpSearchCategoryRepository.findByCpSearchCategorySeq(cpSearchCategorySeq);
+            ResponseCpSearchCategoryDTO dto = cpSearchCategoryRepository.findByCpSearchCategorySeq(cpSearchCategorySeq)
+                    .orElseThrow(() -> {
+                        logger.warn("CP 검색 카테고리를 찾을 수 없습니다: cpSearchCategorySeq={}", cpSearchCategorySeq);
+                        return new CustomException(ErrorCode.NOT_FOUND_CP_SEARCH_CATEGORY);
+                    });
+
             logger.info("CP 검색 카테고리 조회 성공: {}", dto);
+            return dto;
+
         } catch (DataAccessException e) {
             logger.error("데이터 접근 오류 발생: {}", e.getMessage(), e);
             throw new CustomException(ErrorCode.INTERNAL_DATA_ACCESS_ERROR);
         } catch (Exception e) {
             logger.error("예기치 않은 오류 발생: {}", e.getMessage(), e);
-            throw new CustomException(ErrorCode.INTERNAL_SERVER_ERROR);
+            throw new RuntimeException("CP 검색 카테고리 상세 조회 과정에서 예상치 못한 에러가 발생했습니다.", e);
         }
-
-        // 결과가 없을 경우 예외 처리
-        if (dto == null) {
-            logger.warn("CP 검색 카테고리 조회 결과가 없음: cpSearchCategorySeq={}", cpSearchCategorySeq);
-            throw new CustomException(ErrorCode.NOT_FOUND_CP_SEARCH_CATEGORY);
-        }
-
-        return dto;
     }
 
     /**
-     * 새로운 CP 검색 카테고리를 등록합니다.
+     * 새로운 CP 검색 카테고리를 생성하는 메서드입니다.
+     *
+     * <p>주어진 카테고리 이름의 유효성을 검사하고 중복 여부를 확인한 후,
+     * 새로운 CP 검색 카테고리를 데이터베이스에 저장합니다.</p>
      *
      * @param cpSearchCategoryName 생성할 CP 검색 카테고리의 이름
-     * @return 생성된 CP 검색 카테고리의 DTO
-     * @throws CustomException 유효성 검사 실패, 중복 카테고리 또는 데이터베이스 오류가 발생할 경우
+     * @return ResponseCpSearchCategoryDTO 생성된 CP 검색 카테고리의 정보
+     * @throws CustomException 유효성 검사 실패, 중복 카테고리, 또는 데이터 저장 오류 발생 시 예외
      */
     public ResponseCpSearchCategoryDTO createCpSearchCategory(String cpSearchCategoryName) {
         // 1. 유효성 검사 및 중복 검사
-        validateAndCheckDuplicate(cpSearchCategoryName);
+        cpSearchCategoryName = validateAndCheckDuplicate(cpSearchCategoryName);
 
-        // 2. 데이터 저장
+        // 2. 데이터 생성
         CpSearchCategory entity = CpSearchCategory.builder()
                 .userSeq(SecurityUtil.getCurrentUserSeq())
                 .cpSearchCategoryName(cpSearchCategoryName)
                 .build();
 
         try {
+            // 3. 데이터 저장
             entity = cpSearchCategoryRepository.save(entity); // 저장 후 반환된 엔티티를 업데이트
         } catch (DataAccessException e) {
-            logger.error("CP 검색 카테고리 저장 중 오류 발생: {}", e.getMessage());
+            logger.error("CP 검색 카테고리 저장 중 오류 발생: {}", e.getMessage(), e);
             throw new CustomException(ErrorCode.INTERNAL_DATABASE_ERROR);
         }
 
-        // 3. 데이터 반환
-        return cpSearchCategoryRepository.findByCpSearchCategorySeq(entity.getCpSearchCategorySeq());
+        // 4. 저장된 데이터 반환
+        CpSearchCategory finalEntity = entity;
+        return cpSearchCategoryRepository.findByCpSearchCategorySeq(entity.getCpSearchCategorySeq())
+                .orElseThrow(() -> {
+                    logger.warn("저장된 CP 검색 카테고리를 찾을 수 없습니다: cpSearchCategorySeq={}", finalEntity.getCpSearchCategorySeq());
+                    return new CustomException(ErrorCode.NOT_FOUND_CP_SEARCH_CATEGORY);
+                });
     }
 
     /**
-     * CP 검색 카테고리 이름의 유효성을 검사하고 중복 여부를 확인합니다.
+     * 주어진 CP 검색 카테고리 이름의 유효성을 검사하고 중복 여부를 확인합니다.
+     *
+     * <p>이 메서드는 카테고리 이름이 null이거나 비어 있는지 검사하고,
+     * 중복된 이름이 존재할 경우 예외를 발생시킵니다.</p>
      *
      * @param cpSearchCategoryName 확인할 검색 카테고리 이름
+     * @return 유효성 검사를 통과한 검색 카테고리 이름 (공백 제거)
      * @throws CustomException 유효성 검사 실패 또는 중복 카테고리인 경우 발생
      */
-    public void validateAndCheckDuplicate(String cpSearchCategoryName) {
+    public String validateAndCheckDuplicate(String cpSearchCategoryName) {
+
         // 유효성 검사
         if (cpSearchCategoryName == null || cpSearchCategoryName.trim().isEmpty()) {
             throw new CustomException(ErrorCode.REQUIRED_FIELD_MISSING);
@@ -144,6 +151,8 @@ public class CpSearchCategoryService {
         if (existsByName(cpSearchCategoryName)) {
             throw new CustomException(ErrorCode.DUPLICATE_CP_SEARCH_CATEGORY_NAME);
         }
+
+        return cpSearchCategoryName.trim();
     }
 
     /**
@@ -156,19 +165,21 @@ public class CpSearchCategoryService {
         return cpSearchCategoryRepository.findByCpSearchCategoryName(cpSearchCategoryName).isPresent();
     }
 
-
     /**
      * CP 검색 카테고리를 업데이트합니다.
      *
+     * <p>주어진 ID에 해당하는 CP 검색 카테고리를 찾아 이름을 업데이트합니다.
+     * 유효성 검사와 중복 확인을 수행하며, 업데이트된 카테고리의 정보를 반환합니다.</p>
+     *
      * @param cpSearchCategorySeq  수정할 CP 검색 카테고리의 ID
      * @param cpSearchCategoryName 새로운 CP 검색 카테고리 이름
-     * @return 업데이트된 CP 검색 카테고리의 DTO
+     * @return ResponseCpSearchCategoryDTO 업데이트된 CP 검색 카테고리의 DTO
      * @throws CustomException 유효성 검사 실패, 카테고리 없음, 또는 데이터베이스 오류가 발생할 경우
      */
     public ResponseCpSearchCategoryDTO updateCpSearchCategory(long cpSearchCategorySeq, String cpSearchCategoryName) {
         // 1. 유효성 검사 및 중복 확인
-        validateAndCheckDuplicate(cpSearchCategoryName);
         logger.info("CP 검색 카테고리 업데이트 요청: ID={} 이름={}", cpSearchCategorySeq, cpSearchCategoryName);
+        cpSearchCategoryName = validateAndCheckDuplicate(cpSearchCategoryName);
 
         // 2. 수정
         CpSearchCategory entity = cpSearchCategoryRepository.findById(cpSearchCategorySeq)
@@ -185,13 +196,21 @@ public class CpSearchCategoryService {
             cpSearchCategoryRepository.save(entity); // 수정된 엔티티를 데이터베이스에 저장
             logger.info("CP 검색 카테고리 업데이트 성공: ID={}", cpSearchCategorySeq);
         } catch (DataAccessException e) {
-            logger.error("CP 검색 카테고리 업데이트 중 오류 발생: {}", e.getMessage());
+            logger.error("CP 검색 카테고리 업데이트 중 오류 발생: {}", e.getMessage(), e);
             throw new CustomException(ErrorCode.INTERNAL_DATABASE_ERROR);
+        } catch (Exception e) {
+            logger.error("CP 검색 카테고리 업데이트 중 오류 발생: {}", e.getMessage(), e);
+            throw new RuntimeException("CP 검색 카테고리 업데이트 중 예상치 못한 오류 발생", e);
         }
 
         // 4. 반환
-        return cpSearchCategoryRepository.findByCpSearchCategorySeq(cpSearchCategorySeq);
+        return cpSearchCategoryRepository.findByCpSearchCategorySeq(cpSearchCategorySeq)
+                .orElseThrow(() -> {
+                    logger.warn("업데이트된 CP 검색 카테고리를 찾을 수 없습니다: ID={}", cpSearchCategorySeq);
+                    return new CustomException(ErrorCode.NOT_FOUND_CP_SEARCH_CATEGORY);
+                });
     }
+
 
     /**
      * CP 검색 카테고리 엔티티의 데이터를 업데이트합니다.
@@ -239,206 +258,6 @@ public class CpSearchCategoryService {
             logger.info("CP 검색 카테고리 삭제 성공: ID={}", cpSearchCategorySeq);
         } catch (DataAccessException e) {
             logger.error("CP 검색 카테고리 삭제 중 오류 발생: {}", e.getMessage());
-            throw new CustomException(ErrorCode.INTERNAL_DATABASE_ERROR);
-        }
-    }
-
-
-    /**
-     * 주어진 CP 검색 카테고리 ID에 대한 모든 데이터를 조회하여 리스트로 반환합니다.
-     *
-     * @param cpSearchCategorySeq 조회할 CP 검색 카테고리의 ID
-     * @return CP 검색 카테고리 데이터 DTO 리스트
-     * @throws CustomException NOT_FOUND_CP_SEARCH_CATEGORY_DATA:
-     *                         조회 결과가 없을 경우 발생.
-     */
-    public List<ResponseCpSearchCategoryDataDTO> getCpSearchCategoryDataListByCpSearchCategorySeq(long cpSearchCategorySeq) {
-        logger.info("CP 검색 카테고리 데이터 조회 요청 시작: ID={}", cpSearchCategorySeq);
-
-        List<ResponseCpSearchCategoryDataDTO> dtoList;
-
-        // DB 조회
-        try {
-            dtoList = cpSearchCategoryDataRepository.findByCpSearchCategorySeq(cpSearchCategorySeq); // 데이터 조회
-
-            if (dtoList.isEmpty()) {
-                logger.warn("CP 검색 카테고리 데이터 조회 결과가 비어 있습니다: ID={}", cpSearchCategorySeq);
-                throw new CustomException(ErrorCode.NOT_FOUND_CP_SEARCH_CATEGORY_DATA);
-            }
-
-            logger.info("CP 검색 카테고리 데이터 조회 성공: ID={} 데이터 수={}", cpSearchCategorySeq, dtoList.size());
-        } catch (DataAccessException e) {
-            logger.error("데이터 접근 오류 발생: {}", e.getMessage(), e);
-            throw new CustomException(ErrorCode.INTERNAL_DATA_ACCESS_ERROR);
-        } catch (Exception e) {
-            logger.error("예기치 않은 오류 발생: {}", e.getMessage(), e);
-            throw new CustomException(ErrorCode.INTERNAL_SERVER_ERROR);
-        }
-
-        return dtoList;
-    }
-
-    /**
-     * CP 검색 카테고리 데이터를 조회하는 서비스 메서드.
-     *
-     * @param cpSearchCategorySeq     조회할 CP 검색 카테고리의 고유 번호
-     * @param cpSearchCategoryDataSeq 조회할 CP 검색 카테고리 데이터의 고유 번호
-     * @return ResponseCpSearchCategoryDataDTO 해당 데이터의 정보
-     * @throws CustomException 내부 데이터 접근 오류 또는 데이터 미발견 시 예외 발생
-     */
-    public ResponseCpSearchCategoryDataDTO getCpSearchCategoryData(long cpSearchCategorySeq, long cpSearchCategoryDataSeq) {
-        logger.info("CP 검색 카테고리 데이터 조회 시작: cpSearchCategorySeq={}, cpSearchCategoryDataSeq={}", cpSearchCategorySeq, cpSearchCategoryDataSeq);
-
-        ResponseCpSearchCategoryDataDTO dto;
-
-        try {
-            // 데이터 조회
-            dto = cpSearchCategoryDataRepository.findByCpSearchCategoryDataSeq(cpSearchCategoryDataSeq);
-
-            // 결과가 없을 경우 예외 처리
-            if (dto == null) {
-                logger.warn("CP 검색 카테고리 데이터 조회 결과가 없음: cpSearchCategorySeq={}, cpSearchCategoryDataSeq={}", cpSearchCategorySeq, cpSearchCategoryDataSeq);
-                throw new CustomException(ErrorCode.NOT_FOUND_CP_SEARCH_CATEGORY_DATA);
-            }
-
-            logger.info("CP 검색 카테고리 데이터 조회 성공: {}", dto);
-        } catch (DataAccessException e) {
-            logger.error("데이터 접근 오류 발생: {}", e.getMessage(), e);
-            throw new CustomException(ErrorCode.INTERNAL_DATA_ACCESS_ERROR);
-        } catch (Exception e) {
-            logger.error("예기치 않은 오류 발생: {}", e.getMessage(), e);
-            throw new CustomException(ErrorCode.INTERNAL_SERVER_ERROR);
-        }
-
-        return dto;
-    }
-
-    /**
-     * 새로운 CP 검색 카테고리 데이터를 등록합니다.
-     *
-     * @param cpSearchCategorySeq      생성할 CP 검색 카테고리의 시퀀스
-     * @param cpSearchCategoryDataName 생성할 CP 검색 카테고리 데이터의 이름
-     * @return 생성된 CP 검색 카테고리 데이터의 DTO
-     * @throws CustomException 유효성 검사 실패, 중복 데이터 또는 데이터베이스 오류가 발생할 경우
-     */
-    public ResponseCpSearchCategoryDataDTO createCpSearchCategoryData(long cpSearchCategorySeq, String cpSearchCategoryDataName) {
-        // 1. 유효성 검사 및 중복 검사
-        validateAndCheckDuplicateData(cpSearchCategorySeq, cpSearchCategoryDataName);
-
-        // 2. 데이터 저장
-        CpSearchCategoryData entity = CpSearchCategoryData.builder()
-                .userSeq(SecurityUtil.getCurrentUserSeq())
-                .cpSearchCategorySeq(cpSearchCategorySeq)
-                .cpSearchCategoryDataName(cpSearchCategoryDataName)
-                .build();
-
-        try {
-            entity = cpSearchCategoryDataRepository.save(entity); // 저장 후 반환된 엔티티를 업데이트
-        } catch (DataAccessException e) {
-            logger.error("CP 검색 카테고리 데이터 저장 중 오류 발생: {}", e.getMessage());
-            throw new CustomException(ErrorCode.INTERNAL_DATABASE_ERROR);
-        }
-
-        // 3. 데이터 반환
-        return cpSearchCategoryDataRepository.findByCpSearchCategoryDataSeq(entity.getCpSearchCategoryDataSeq());
-    }
-
-    /**
-     * CP 검색 카테고리 데이터의 유효성을 검사하고 중복 여부를 확인합니다.
-     *
-     * @param cpSearchCategorySeq      생성할 CP 검색 카테고리의 시퀀스
-     * @param cpSearchCategoryDataName 생성할 CP 검색 카테고리 데이터의 이름
-     * @throws CustomException 유효성 검사 실패 시 REQUIRED_FIELD_MISSING 예외 발생,
-     *                         중복 데이터가 발견될 경우 DUPLICATE_CP_SEARCH_CATEGORY_DATA_NAME 예외 발생
-     */
-    private void validateAndCheckDuplicateData(long cpSearchCategorySeq, String cpSearchCategoryDataName) {
-        // 유효성 검사 로직 (예: null 체크)
-        if (cpSearchCategoryDataName == null || cpSearchCategoryDataName.trim().isEmpty()) {
-            throw new CustomException(ErrorCode.REQUIRED_FIELD_MISSING);
-        }
-
-        // 중복 검사 로직
-        boolean exists = cpSearchCategoryDataRepository.existsByCpSearchCategoryDataName(cpSearchCategoryDataName);
-        if (exists) {
-            throw new CustomException(ErrorCode.DUPLICATE_CP_SEARCH_CATEGORY_DATA_NAME);
-        }
-    }
-
-    /**
-     * CP 검색 카테고리 데이터를 업데이트합니다.
-     *
-     * @param cpSearchCategoryDataSeq 수정할 CP 검색 카테고리 데이터의 ID
-     * @param cpSearchCategoryDataName 새로운 CP 검색 카테고리 데이터 이름
-     * @return 업데이트된 CP 검색 카테고리 데이터의 DTO
-     * @throws CustomException 유효성 검사 실패, 데이터 없음, 또는 데이터베이스 오류가 발생할 경우
-     */
-    public ResponseCpSearchCategoryDataDTO updateCpSearchCategoryDataData(long cpSearchCategoryDataSeq, String cpSearchCategoryDataName) {
-        // 1. 유효성 검사
-        validateAndCheckDuplicateData(cpSearchCategoryDataSeq, cpSearchCategoryDataName);
-        logger.info("CP 검색 카테고리 데이터 업데이트 요청: ID={} 이름={}", cpSearchCategoryDataSeq, cpSearchCategoryDataName);
-
-        // 2. 수정할 데이터 찾기
-        CpSearchCategoryData entity = cpSearchCategoryDataRepository.findById(cpSearchCategoryDataSeq)
-                .orElseThrow(() -> {
-                    logger.warn("CP 검색 카테고리 데이터 찾기 실패: ID={}", cpSearchCategoryDataSeq);
-                    return new CustomException(ErrorCode.NOT_FOUND_CP_SEARCH_CATEGORY_DATA);
-                });
-
-        // 카테고리 데이터 업데이트
-        updateCpSearchCategoryData(entity, cpSearchCategoryDataName);
-
-        // 3. 데이터 저장
-        try {
-            cpSearchCategoryDataRepository.save(entity); // 수정된 엔티티를 데이터베이스에 저장
-            logger.info("CP 검색 카테고리 데이터 업데이트 성공: ID={}", cpSearchCategoryDataSeq);
-        } catch (DataAccessException e) {
-            logger.error("CP 검색 카테고리 데이터 업데이트 중 오류 발생: {}", e.getMessage());
-            throw new CustomException(ErrorCode.INTERNAL_DATABASE_ERROR);
-        }
-
-        // 4. 반환
-        return cpSearchCategoryDataRepository.findByCpSearchCategoryDataSeq(cpSearchCategoryDataSeq);
-    }
-
-    /**
-     * CP 검색 카테고리 데이터 엔티티의 데이터를 업데이트합니다.
-     *
-     * @param entity                    수정할 CP 검색 카테고리 데이터 엔티티
-     * @param cpSearchCategoryDataName  새로운 CP 검색 카테고리 데이터 이름
-     */
-    public void updateCpSearchCategoryData(CpSearchCategoryData entity, String cpSearchCategoryDataName) {
-        entity.updateUserSeq(SecurityUtil.getCurrentUserSeq()); // 현재 사용자 ID 업데이트
-        entity.updateCpSearchCategoryDataName(cpSearchCategoryDataName); // 데이터 이름 업데이트
-        logger.info("CP 검색 카테고리 데이터 업데이트: ID={} 이름={}", entity.getCpSearchCategoryDataSeq(), cpSearchCategoryDataName);
-    }
-
-    /**
-     * CP 검색 카테고리 데이터를 삭제합니다.
-     *
-     * @param cpSearchCategoryDataSeq 삭제할 CP 검색 카테고리 데이터의 ID
-     * @throws CustomException 데이터를 찾지 못한 경우 발생
-     */
-    public void deleteCpSearchCategoryData(long cpSearchCategoryDataSeq) {
-        // 1. 조회
-        CpSearchCategoryData entity = cpSearchCategoryDataRepository.findById(cpSearchCategoryDataSeq)
-                .orElseThrow(() -> {
-                    logger.warn("CP 검색 카테고리 데이터 찾기 실패: ID={}", cpSearchCategoryDataSeq);
-                    return new CustomException(ErrorCode.NOT_FOUND_CP_SEARCH_CATEGORY_DATA);
-                });
-
-        // 2. CP 검색 카테고리 데이터 삭제 요청
-        logger.info("CP 검색 카테고리 데이터 삭제 요청: ID={}", cpSearchCategoryDataSeq);
-
-        // 3. 삭제(삭제일 추가)
-        entity.updateUserSeq(SecurityUtil.getCurrentUserSeq()); // 현재 사용자 ID 업데이트
-        entity.delete(); // 삭제 처리
-
-        // 4. 저장
-        try {
-            cpSearchCategoryDataRepository.save(entity); // 수정된 엔티티를 데이터베이스에 저장
-            logger.info("CP 검색 카테고리 데이터 삭제 성공: ID={}", cpSearchCategoryDataSeq);
-        } catch (DataAccessException e) {
-            logger.error("CP 검색 카테고리 데이터 삭제 중 오류 발생: {}", e.getMessage());
             throw new CustomException(ErrorCode.INTERNAL_DATABASE_ERROR);
         }
     }
